+++ components/brave_extension/extension/brave_extension/braveShieldsPanel.html
@@ -3,7 +3,7 @@
     <head>
         <meta charset="utf-8">
         <meta name="viewport" content="width=device-width">
-        <link rel="stylesheet" href="/bravelizer.css">
+        <link rel="stylesheet" href="/assets/bravelizer.css">
         <title>Brave Shields Panel</title>
         <style>
             body {
@@ -13,6 +13,6 @@
     </head>
     <body>
         <div id="root"></div>
-        <script src="/brave_extension.bundle.js"></script>
+        <script src="/out/brave_extension.bundle.js"></script>
     </body>
 </html>
