+++ components/brave_extension/extension/brave_extension/package.json
@@ -8,7 +8,7 @@
     "compress": "node scripts/compress",
     "compress-keygen": "crx keygen",
     "clean": "rimraf build/ dev/ *.zip *.crx",
-    "lint": "tslint --project tsconfig.json '{app, test}/**/*.{ts,tsx}'",
+    "lint": "tslint --project tsconfig.json '{test}/**/*.{ts,tsx}'",
     "test-unit": "cross-env NODE_ENV=test TS_NODE_TYPE_CHECK=true jest --config=./test/jest.config.json",
     "test-security": "npm audit",
     "test-suite": "node ./scripts/tests.js"
