+++ components/brave_extension/extension/brave_extension/test/jest.config.json
@@ -13,15 +13,15 @@
   "resetModules": true,
   "collectCoverage": false,
   "testMatch": [
-    "<rootDir>/app/actions/*.ts",
-    "<rootDir>/app/background/api/*.ts",
-    "<rootDir>/app/components/**/*.tsx",
-    "<rootDir>/app/helpers/*.tsx",
-    "<rootDir>/app/background/reducersTest.ts",
-    "<rootDir>/app/background/storeTest.ts",
-    "<rootDir>/app/background/reducers/*.ts",
-    "<rootDir>/app/state/*.ts",
-    "<rootDir>/app/background/events/*.ts"
+    "<rootDir>/actions/*.ts",
+    "<rootDir>/background/api/*.ts",
+    "<rootDir>/components/**/*.tsx",
+    "<rootDir>/helpers/*.tsx",
+    "<rootDir>/background/reducersTest.ts",
+    "<rootDir>/background/storeTest.ts",
+    "<rootDir>/background/reducers/*.ts",
+    "<rootDir>/state/*.ts",
+    "<rootDir>/background/events/*.ts"
   ],
   "transformIgnorePatterns": [
     "<rootDir>/node_modules"
